delete fix profil lulusan

+++ app/Http/Controllers/ProfilLulusanController.php
@@ -6,6 +6,7 @@ use App\Exports\ProfilLulusanExport;
 use Dompdf\Dompdf;
 use Illuminate\Http\Request;
 use App\Models\Profil_Lulusan;
+use App\Models\Detail_PL_CPLProdi;
 use Maatwebsite\Excel\Facades\Excel;
 use Illuminate\Support\Facades\Response;
 use Illuminate\Support\Facades\Validator;
@@ -111,9 +112,13 @@ class ProfilLulusanController extends Controller
 
     public function delete($pl)
     {
-        $pl = Profil_Lulusan::where('kodePL', $pl)->first();
-        $pl->delete();
+        if (Detail_PL_CPLProdi::where('kodePL', $pl)->exists()) {
+            return redirect()->route('kurikulum.data.profil_lulusan')->with('error', 'Profil Lulusan masih berelasi dengan CPL Prodi.');
+        }else{
+            $pl = Profil_Lulusan::where('kodePL', $pl)->first();
+            $pl->delete();
+            return redirect()->route('kurikulum.data.profil_lulusan')->with('success', 'Profil Lulusan berhasil dihapus');
+        }
 
-        return redirect()->route('kurikulum.data.profil_lulusan')->with('success', 'Profil Lulusan berhasil dihapus');
     }
 }
